% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aipw.R
\name{aipw}
\alias{aipw}
\title{AIPW}
\usage{
aipw(exposure, outcome, tmle_fit, tmle3_fit = NULL)
}
\arguments{
\item{exposure}{binary numeric vector}

\item{outcome}{binary numeric vector}

\item{tmle_fit}{fitted tmle object}

\item{tmle3_fit}{place holder for tmle3 function}
}
\value{
a matrix with risk difference (RD), risk ratio (RR), odds ratio (OR)
}
\description{
Augmented Inverse Probability Weighting
}
