% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AIPW_base.R
\docType{class}
\name{AIPW_base}
\alias{AIPW_base}
\title{Augmented Inverse Probablity Weighting Base Class (AIPW_base)}
\format{
\code{\link{R6Class}} object.
}
\value{
\code{AIPW} base object
}
\description{
Define a base R6Class for later subclasses inheritance of
\code{AIPW_base$summary()} and \code{AIPW_base$plot.p_score()}
}
\examples{

## ------------------------------------------------
## Method `AIPW_base$summary`
## ------------------------------------------------

library(SuperLearner)
aipw_sl <- AIPW$new(Y=rbinom(100,1,0.5), A=rbinom(100,1,0.5),
                    W.Q=rbinom(100,1,0.5), W.g=rbinom(100,1,0.5),
                    Q.SL.library="SL.mean",g.SL.library="SL.mean",
                    k_split=1,verbose=FALSE)$fit()
aipw_sl$summary(g.bound=0.025)

## ------------------------------------------------
## Method `AIPW_base$plot.p_score`
## ------------------------------------------------

library(SuperLearner)
library(ggplot2)
aipw_sl <- AIPW$new(Y=rbinom(100,1,0.5), A=rbinom(100,1,0.5),
                    W.Q=rbinom(100,1,0.5), W.g=rbinom(100,1,0.5),
                    Q.SL.library="SL.mean",g.SL.library="SL.mean",
                    k_split=1,verbose=FALSE)$fit()
#before average treatment effect calculation
aipw_sl$plot.p_score()
#after calculation
aipw_sl$summary(g.bound=0.025)$plot.p_score()
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{n}}{Number of observations}

\item{\code{obs_est}}{Components for estimating the influence functions of all observations to calculate average causal effects}

\item{\code{estimates}}{Risk difference, risk ratio, odds ratio and variance-covariance matrix for SE calculation}

\item{\code{result}}{A matrix contains RD, RR and OR with their SE and 95\%CI}

\item{\code{g.plot}}{A density plot of propensity scores by exposure status (\code{ggplot2::geom_density})}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{AIPW_base$new()}}
\item \href{#method-summary}{\code{AIPW_base$summary()}}
\item \href{#method-plot.p_score}{\code{AIPW_base$plot.p_score()}}
\item \href{#method-clone}{\code{AIPW_base$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new \code{AIPW_base} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AIPW_base$new(Y = NULL, A = NULL, verbose = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{Y}}{Outcome (binary integer: 0 or 1)}

\item{\code{A}}{Exposure (binary integer: 0 or 1)}

\item{\code{verbose}}{Whether to print the result (logical; Default = FALSE)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{AIPW_base} obejct
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-summary"></a>}}
\if{latex}{\out{\hypertarget{method-summary}{}}}
\subsection{Method \code{summary()}}{
Calculate average causal effects in RD, RR and OR in the fitted \code{AIPW} obejct with the estimated influence functions
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AIPW_base$summary(g.bound = 0.025)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{g.bound}}{Value between [0,1] at which the propensity score should be truncated. Defaults to 0.025.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An \code{AIPW} object with average treatment effect estimations in RD, RR and OR
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{library(SuperLearner)
aipw_sl <- AIPW$new(Y=rbinom(100,1,0.5), A=rbinom(100,1,0.5),
                    W.Q=rbinom(100,1,0.5), W.g=rbinom(100,1,0.5),
                    Q.SL.library="SL.mean",g.SL.library="SL.mean",
                    k_split=1,verbose=FALSE)$fit()
aipw_sl$summary(g.bound=0.025)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plot.p_score"></a>}}
\if{latex}{\out{\hypertarget{method-plot.p_score}{}}}
\subsection{Method \code{plot.p_score()}}{
Plot and check the balance of propensity scores by exposure status
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AIPW_base$plot.p_score()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{g.plot} A density plot of propensity scores by exposure status (\code{ggplot2::geom_density})
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{library(SuperLearner)
library(ggplot2)
aipw_sl <- AIPW$new(Y=rbinom(100,1,0.5), A=rbinom(100,1,0.5),
                    W.Q=rbinom(100,1,0.5), W.g=rbinom(100,1,0.5),
                    Q.SL.library="SL.mean",g.SL.library="SL.mean",
                    k_split=1,verbose=FALSE)$fit()
#before average treatment effect calculation
aipw_sl$plot.p_score()
#after calculation
aipw_sl$summary(g.bound=0.025)$plot.p_score()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AIPW_base$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
