% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AIPW_base.R
\name{plot.p_score}
\alias{plot.p_score}
\title{Plot the propensity scores by exposure status}
\value{
\code{g.plot} (public variable): A density plot of propensity scores by exposure status (\code{ggplot2::geom_density})
}
\description{
Plot and check the balance of propensity scores by exposure status
}
\section{R6 Usage}{

\code{$plot.p_plot()}
}

\examples{
library(SuperLearner)
library(ggplot2)
aipw_sl <- AIPW$new(Y=rbinom(100,1,0.5), A=rbinom(100,1,0.5),
                    W.Q=rbinom(100,1,0.5), W.g=rbinom(100,1,0.5),
                    Q.SL.library="SL.mean",g.SL.library="SL.mean",
                    k_split=1,verbose=FALSE)$fit()
#before average treatment effect calculation
aipw_sl$plot.p_score()
#after calculation
aipw_sl$summary(g.bound=0.025)$plot.p_score()
}
\seealso{
\link{AIPW} and \link{AIPW_tmle}
}
