% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AIPW.R
\name{stratified_fit}
\alias{stratified_fit}
\alias{stratified_fit.AIPW}
\title{Fit the data to the \link{AIPW} object stratified by \code{A} for the outcome model}
\value{
A fitted \link{AIPW} object with \code{obs_est} and \code{libs} (public variables)
}
\description{
Fitting the data into the \link{AIPW} object with/without cross-fitting to estimate the efficient influence functions.
Outcome model is fitted, stratified by exposure status \code{A}
}
\section{R6 Usage}{

\code{$stratified_fit.AIPW()}
}

\examples{
library(SuperLearner)
aipw_sl <- AIPW$new(Y=rbinom(100,1,0.5), A=rbinom(100,1,0.5),
                    W.Q=rbinom(100,1,0.5), W.g=rbinom(100,1,0.5),
                    Q.SL.library="SL.mean",g.SL.library="SL.mean",
                    k_split=1,verbose=FALSE)
aipw_sl$stratified_fit()
}
\seealso{
\link{AIPW}
}
