% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Repeated.R
\name{summary_median}
\alias{summary_median}
\alias{summary_median.Repeated}
\title{Summary of the \code{repeated_estimates} from \code{repfit()} in the \link{Repeated} object using median methods.}
\value{
\code{repeated_results} and \code{result} (public variables).
}
\description{
From \code{repeated_estimates}, calculate the median estimate (\code{median(Estimates)}), median SE (\code{median(SE)}), SE adjusting for variations across \code{num_reps} times,
and 95\% CI using SE adjusting for SE adjusted for variability.
}
\section{R6 Usage}{

\code{$summary_median.Repeated()}
}

\references{
Chernozhukov V, Chetverikov V, Demirer M, et al (2018). Double/debiased machine learning for treatment and structural parameters. \emph{The Econometrics Journal}.
}
\seealso{
\link{Repeated} and \link{AIPW}
}
