% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aipw.R
\name{aipw}
\alias{aipw}
\title{AIPW}
\usage{
aipw(aipw_input = NULL, tmle_fit = NULL, A = NULL, Y = NULL)
}
\arguments{
\item{aipw_input}{a matrix of individual estimates created by  \code{\link[=aipw_input]{aipw_input()}}}

\item{tmle_fit}{fitted `tmle`` object}

\item{A}{exposure (only needed when using tmle fitted object as input)}

\item{Y}{outcome (only needed when using tmle fitted object as input)}
}
\value{
a matrix with risk difference (RD), risk ratio (RR), odds ratio (OR)
}
\description{
Augmented Inverse Probability Weighting
}
