% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AIPW.R
\docType{class}
\name{AIPW}
\alias{AIPW}
\title{Augmented Inverse Probablity Weighting (AIPW)}
\format{
\code{\link{R6Class}} object.
}
\value{
\code{AIPW} object
}
\description{
Define an R6Class aipw object
}
\details{
create an AIPW object
}
\examples{

## ------------------------------------------------
## Method `AIPW$new`
## ------------------------------------------------

library(SuperLearner)
aipw_sl <- AIPW$new(Y=rbinom(100,1,0.5), A=rbinom(100,1,0.5),
                    W.Q=rbinom(100,1,0.5), W.g=rbinom(100,1,0.5),
                    Q.SL.library="SL.mean",g.SL.library="SL.mean",
                    k_split=1,verbose=FALSE)

## ------------------------------------------------
## Method `AIPW$fit`
## ------------------------------------------------

library(SuperLearner)
aipw_sl <- AIPW$new(Y=rbinom(100,1,0.5), A=rbinom(100,1,0.5),
                    W.Q=rbinom(100,1,0.5), W.g=rbinom(100,1,0.5),
                    Q.SL.library="SL.mean",g.SL.library="SL.mean",
                    k_split=1,verbose=FALSE)
aipw_sl$fit()
}
\section{Super class}{
\code{\link[AIPW:AIPW_base]{AIPW::AIPW_base}} -> \code{AIPW}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{libs}}{SuperLearner or sl3 libraries and their fitted objects}

\item{\code{sl.fit}}{a wrapper for fitting SuperLearner or sl3}

\item{\code{sl.predict}}{a wrapper using \code{sl.fit} to predict}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{AIPW$new()}}
\item \href{#method-fit}{\code{AIPW$fit()}}
\item \href{#method-clone}{\code{AIPW$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="AIPW" data-topic="AIPW_base" data-id="calculate_result">}\href{../../AIPW/html/AIPW_base.html#method-calculate_result}{\code{AIPW::AIPW_base$calculate_result()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="AIPW" data-topic="AIPW_base" data-id="plot.p_score">}\href{../../AIPW/html/AIPW_base.html#method-plot.p_score}{\code{AIPW::AIPW_base$plot.p_score()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new \code{AIPW} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AIPW$new(
  Y = NULL,
  A = NULL,
  verbose = FALSE,
  W.Q = NULL,
  W.g = NULL,
  Q.SL.library = NULL,
  g.SL.library = NULL,
  k_split = 1
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{Y}}{outcome (binary integer: 0 or 1)}

\item{\code{A}}{exposure (binary integer: 0 or 1)}

\item{\code{verbose}}{whether to show progression bar and print the result (logical; Default = FALSE)}

\item{\code{W.Q}}{covariates for outcome model (vector, matrix or data.frame)}

\item{\code{W.g}}{covariates for exposure model (vector, matrix or data.frame)}

\item{\code{Q.SL.library}}{SuperLearner libraries or sl3 learner object (Lrnr_base) for outcome model}

\item{\code{g.SL.library}}{SuperLearner libraries or sl3 learner object (Lrnr_base) for exposure model}

\item{\code{k_split}}{number of splitting (integer; range: from 1 to number of observation-1):
if k_split=1, no sample splitting;
if k_split>1, use similar technique of cross-validation
(e.g., k_split=5, use 4/5 of the data to estimate and the remaining 1/5 leftover to predict)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{AIPW} obejct
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{library(SuperLearner)
aipw_sl <- AIPW$new(Y=rbinom(100,1,0.5), A=rbinom(100,1,0.5),
                    W.Q=rbinom(100,1,0.5), W.g=rbinom(100,1,0.5),
                    Q.SL.library="SL.mean",g.SL.library="SL.mean",
                    k_split=1,verbose=FALSE)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fit"></a>}}
\if{latex}{\out{\hypertarget{method-fit}{}}}
\subsection{Method \code{fit()}}{
Fitting the data into the \code{AIPW} object with/without sample splitting to estimate the influence functions
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AIPW$fit()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A fitted \code{AIPW} obejct
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{library(SuperLearner)
aipw_sl <- AIPW$new(Y=rbinom(100,1,0.5), A=rbinom(100,1,0.5),
                    W.Q=rbinom(100,1,0.5), W.g=rbinom(100,1,0.5),
                    Q.SL.library="SL.mean",g.SL.library="SL.mean",
                    k_split=1,verbose=FALSE)
aipw_sl$fit()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AIPW$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
