% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AIPW.R
\name{AIPW}
\alias{AIPW}
\title{Augmented Inverse Probability Weighting (AIPW)}
\value{
\code{AIPW} object
}
\description{
An R6Class of AIPW for estimating the average causal effects with users' inputs of exposure, outcome, covariates and related
libraries for estimating the efficient influence function.
}
\details{
An AIPW object is constructed by \code{new()} with users' inputs of data and causal structures, then it \code{fit()} the data using the
libraries in \code{Q.SL.library} and \code{g.SL.library} with \code{k_split} sample splitting, and provides results via the \code{summary()} method.
After using \code{fit()} and/or \code{summary()} methods, the propensity scores by exposure status can be examined with \code{plot.p_score()}.
See examples for illustration.
}
\section{Constructor}{

\code{AIPW$new(Y = NULL, A = NULL, W = NULL, W.Q = NULL, W.g = NULL, Q.SL.library = NULL, g.SL.library = NULL, k_split = 10, verbose = TRUE)}
}

\section{Constructor Arguments}{

\tabular{lll}{
\strong{Argument}      \tab   \strong{Type}     \tab     \strong{Details} \cr
\code{Y}               \tab   Integer           \tab     A vector of outcome (0 or 1) \cr
\code{A}               \tab   Integer           \tab     A vector of exposure (0 or 1) \cr
\code{W}               \tab   Data              \tab    Vector, matrix or data.frame of covariates for both exposure and outcome models.
If NULL, this function will seek for inputs from \code{W.Q} and \code{W.g}. \cr
\code{W.Q}             \tab   Data              \tab    Vector, matrix or data.frame of covariates for the outcome model (Q).
Only valid when \code{W} is NULL, otherwise it would be replaced by \code{W}. \cr
\code{W.g}             \tab   Data              \tab    Vector, matrix or data.frame of covariates for the exposure model (g).
Only valid when \code{W} is NULL, otherwise it would be replaced by \code{W}. \cr
\code{Q.SL.library}    \tab   SL.library        \tab    \link{SuperLearner} libraries or \code{sl3} learner object (Lrnr_base) of algorithms used for the outcome model (Q). \cr
\code{g.SL.library}    \tab   SL.library        \tab    \link{SuperLearner} libraries or \code{sl3} learner object (Lrnr_base) of algorithms used for the exposure model (g). \cr
\code{k_split}         \tab   Integer           \tab    Number of splitting (Default = 10; range: from 1 to number of observation-1).
If k_split=1, no sample splitting; if k_split>1, use similar technique as cross-validation
(e.g., k_split=10, use 9/10 of the data to estimate and the remaining 1/10 leftover to predict
\strong{NOTE: it's recommended to use sample splitting.} \cr
\code{verbose}         \tab   Logical           \tab    Whether to print the result (Default = TRUE) \cr
}
}

\section{Public Methods}{

\tabular{lll}{
\strong{Methods}      \tab   \strong{Details}                                      \tab \strong{Link}     \cr
\code{fit()}          \tab   Fit the data to the \link{AIPW} object                     \tab   \link{fit.AIPW}  \cr
\code{summary()}      \tab   Summary of the average treatment effects from AIPW    \tab   \link{summary.AIPW_base}\cr
\code{plot.p_score()} \tab   Plot the propensity scores by exposure status         \tab   \link{plot.p_score}\cr
}
}

\section{Public Variables}{

\tabular{lll}{
\strong{Variable}     \tab   \strong{Generated by}      \tab     \strong{Return} \cr
\code{n}              \tab   Constructor                \tab     Number of observations \cr
\code{obs_est}        \tab   \code{fit()}                    \tab     Components calculating average causal effects,
including propensity scores, counterfactual predictions and efficient influence functions \cr
\code{estimates}      \tab   \code{summary()}                \tab     Risk difference, risk ratio, odds ratio and variance-covariance matrix for SE calculation \cr
\code{result}         \tab   \code{summary()}                \tab     A matrix contains RD, RR and OR with their SE and 95\%CI \cr
\code{g.plot}         \tab   \code{plot.p_score()}           \tab     A density plot of propensity scores by exposure status (\code{ggplot2::geom_density})\cr
\code{libs}           \tab   \code{fit()}                    \tab     \link{SuperLearner} or sl3 libraries and their fitted objects \cr
\code{sl.fit}         \tab   Constructor                \tab     A wrapper function for fitting \link{SuperLearner} or sl3 \cr
\code{sl.predict}     \tab   Constructor                \tab     A wrapper function using \code{sl.fit} to predict \cr
}
}

\examples{
library(SuperLearner)
library(ggplot2)

#create an object
aipw_sl <- AIPW$new(Y=rbinom(100,1,0.5), A=rbinom(100,1,0.5),
                    W.Q=rbinom(100,1,0.5), W.g=rbinom(100,1,0.5),
                    Q.SL.library="SL.mean",g.SL.library="SL.mean",
                    k_split=1,verbose=FALSE)$fit()

#fit the object
aipw_sl$fit()

#calculate the retults
aipw_sl$summary(g.bound = 0.025)

#check the propensity scores by exposure status after truncation
aipw_sl$plot.p_score()

}
