% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aipw.R
\docType{class}
\name{aipw}
\alias{aipw}
\title{Augmented Inverse Probablity Weighting (AIPW)

Define an AIPW object}
\format{
\code{\link{R6Class}} object.
}
\value{
\code{aipw} object
}
\description{
Augmented Inverse Probablity Weighting (AIPW)

Define an AIPW object

Augmented Inverse Probablity Weighting (AIPW)

Define an AIPW object
}
\details{
create an AIPW object
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{n}}{number of observations}

\item{\code{libs}}{SuperLearner or sl3 libraries and their fitted objects}

\item{\code{obs_est}}{estimates for each observation to calculate average causal effects}

\item{\code{estimates}}{risk difference, risk ratio, odds ratio and variance-covariance matrix for SE calcualtion}

\item{\code{sl.fit}}{a wrapper for fitting SuperLearner or sl3}

\item{\code{sl.predict}}{a wrapper using \code{sl.fit} to predict}

\item{\code{result}}{a matrix contains RD, RR and OR with their SE and 95\%CI}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{aipw$new()}}
\item \href{#method-calculate_result}{\code{aipw$calculate_result()}}
\item \href{#method-clone}{\code{aipw$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new AIPW object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{aipw$new(
  Y = NULL,
  A = NULL,
  W.Q = NULL,
  W.g = NULL,
  Q.SL.library = NULL,
  g.SL.library = NULL,
  k_split = 1,
  verbose = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{Y}}{outcome}

\item{\code{A}}{exposure}

\item{\code{W.Q}}{covariates for outcome model}

\item{\code{W.g}}{covariates for exposure model}

\item{\code{Q.SL.library}}{SuperLearner libraries or sl3 learner object (Lrnr_base) for outcome model}

\item{\code{g.SL.library}}{SuperLearner libraries or sl3 learner object (Lrnr_base) for exposure model}

\item{\code{k_split}}{number of splitting: if k_split=1, no sample splitting;
if k_split>1, use similar technique of cross-validation
(e.g., k_split=5, use 4/5 of the data to estimate the 1/5 leftover data)}

\item{\code{verbose}}{logical. Default = FALSE; when TRUE, show a text progress bar.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{aipw} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-calculate_result"></a>}}
\if{latex}{\out{\hypertarget{method-calculate_result}{}}}
\subsection{Method \code{calculate_result()}}{
Calculate average causal effects in RD, RR and OR
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{aipw$calculate_result()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{aipw$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
