% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AIPW_base.R
\docType{class}
\name{AIPW_base}
\alias{AIPW_base}
\title{Augmented Inverse Probability Weighting Base Class (AIPW_base)}
\format{
\CRANpkg{R6} object.
}
\value{
\code{AIPW} base object
}
\description{
A base class for AIPW that implements the common methods, such as \code{summary()} and \code{plot.p_score()}, inheritted by \link{AIPW} and \link{AIPW_tmle} class
}
\seealso{
\link{AIPW} and \link{AIPW_tmle}
}
