% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aipw.R
\name{aipw_input}
\alias{aipw_input}
\title{AIPW inputs}
\usage{
aipw_input(
  Y,
  A,
  W.Q,
  W.g,
  Q.SL.library,
  g.SL.library,
  k_split = 1,
  verbose = FALSE
)
}
\arguments{
\item{Y}{a vector of outcome}

\item{A}{a vector of treatment}

\item{W.Q}{a matrix or data.frame of covariates for outcome model}

\item{W.g}{a matrix or data.frame of covariates for exposure model}

\item{Q.SL.library}{outcome model}

\item{g.SL.library}{expousre model (propensity score)}

\item{k_split}{number of splitting: if k_split=1, no sample splitting;
if k_split>1, use similar technique of cross-validation
(e.g., k_split=5, use 4/5 of the data to estimate the 1/5 leftover data)}

\item{verbose}{logical. Default = FALSE; when TRUE, show a text progress bar.}
}
\value{
a list of 1) matrix with estimates for each row when treated aipw_eif1 and untreated aipw_eif0,
2) Q_fit object & 3) g_fit object
}
\description{
Augmented Inverse Probability Weighting
}
\details{
Creating an input for \code{\link[=aipw]{aipw()}}
}
