% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AIPW_tmle.R
\docType{class}
\name{AIPW_tmle}
\alias{AIPW_tmle}
\title{Augmented Inverse Probablity Weighting (AIPW) take TMLE or tmle3 inputs}
\format{
\code{\link{R6Class}} object.
}
\value{
\code{AIPW} object
}
\description{
Define an R6Class aipw_tmle object whcih takes TMLE or tmle3 inputs
}
\details{
create an AIPW object
}
\section{Super class}{
\code{\link[AIPW:AIPW_base]{AIPW::AIPW_base}} -> \code{AIPW_tmle}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{AIPW_tmle$new()}}
\item \href{#method-clone}{\code{AIPW_tmle$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="AIPW" data-topic="AIPW_base" data-id="calculate_result">}\href{../../AIPW/html/AIPW_base.html#method-calculate_result}{\code{AIPW::AIPW_base$calculate_result()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="AIPW" data-topic="AIPW_base" data-id="plot.p_score">}\href{../../AIPW/html/AIPW_base.html#method-plot.p_score}{\code{AIPW::AIPW_base$plot.p_score()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new \code{AIPW_tmle} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AIPW_tmle$new(Y = NULL, A = NULL, tmle_fit = NULL, verbose = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{Y}}{outcome (binary integer: 0 or 1)}

\item{\code{A}}{exposure (binary integer: 0 or 1)}

\item{\code{tmle_fit}}{a fitted \code{tmle} object}

\item{\code{verbose}}{whether to show progression bar and print the result (logical; Default = FALSE)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{AIPW_tmle} obejct
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AIPW_tmle$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
