% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AIPW_base.R
\name{plot.ip_weights}
\alias{plot.ip_weights}
\title{Plot the inverse probability weights using truncated propensity scores by exposure status}
\value{
\code{ip_weights.plot} (public variable): A box plot of inverse probability weights using truncated propensity scores by exposure status (\code{ggplot2::geom_boxplot})
}
\description{
Plot and check the balance of propensity scores by exposure status
}
\section{R6 Usage}{

\code{$plot.ip_weights()}
}

\seealso{
\link{AIPW} and \link{AIPW_tmle}
}
