% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AIPW_base.R
\name{plot.p_score}
\alias{plot.p_score}
\title{Plot the propensity scores by exposure status}
\value{
\code{g.plot} (public variable): A density plot of propensity scores by exposure status (\code{ggplot2::geom_density})
}
\description{
Plot and check the balance of propensity scores by exposure status
}
\section{R6 Usage}{

\code{$plot.p_plot()}
}

\seealso{
\link{AIPW} and \link{AIPW_tmle}
}
