% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AIPW.R
\name{stratified_fit}
\alias{stratified_fit}
\alias{stratified_fit.AIPW}
\title{Fit the data to the \link{AIPW} object stratified by \code{A} for the outcome model}
\value{
A fitted \link{AIPW} object with \code{obs_est} and \code{libs} (public variables)
}
\description{
Fitting the data into the \link{AIPW} object with/without cross-fitting to estimate the efficient influence functions.
Outcome model is fitted, stratified by exposure status \code{A}
}
\section{R6 Usage}{

\code{$stratified_fit.AIPW()}
}

\seealso{
\link{AIPW}
}
