% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AIPW_base.R
\name{summary}
\alias{summary}
\alias{summary.AIPW_base}
\title{Summary of the average treatment effects from AIPW}
\usage{
summary.AIPW_base(object, g.bound = 0.025)
}
\arguments{
\item{object}{An object of \link{AIPW} or \link{AIPW_tmle} class  efficient influence functions}

\item{g.bound}{Value between [0,1] at which the propensity score should be truncated. Defaults to 0.025.}
}
\value{
Summaries of the average treatment effect estimations in RD, RR and OR
}
\description{
Calculate average causal effects in RD, RR and OR in the fitted \code{AIPW} obejct with the estimated influence functions
}
\section{R6 Usage}{

\code{$summary(g.bound = 0.025)}
}

\examples{
library(SuperLearner)
aipw_sl <- AIPW$new(Y=rbinom(100,1,0.5), A=rbinom(100,1,0.5),
                    W.Q=rbinom(100,1,0.5), W.g=rbinom(100,1,0.5),
                    Q.SL.library="SL.mean",g.SL.library="SL.mean",
                    k_split=1,verbose=FALSE)$fit()
aipw_sl$summary(g.bound=0.025)
}
